refactor: move login to pages
